+++ fusion_mlx/watermark/__init__.py
@@ -1,0 +1,15 @@
+from .lsb import (
+    bits_to_payload,
+    compute_signature,
+    embed_bits,
+    extract_bits,
+    payload_to_bits,
+)
+
+__all__ = [
+    "bits_to_payload",
+    "compute_signature",
+    "embed_bits",
+    "extract_bits",
+    "payload_to_bits",
+]